\name{expression}
\docType{data}
\alias{expression}
\title{Expression data.}
\description{
Numeric matrix containing expression data for 23 transcripts (rows) and 41 samples (columns). 

The data correspond to samples 1, 101, 11, 126, 14, 15, 17, 2, 20, 21, 52, 64, 8, H104, H109, H113, H115, H117, H118, H120, H121, H124, H126, H128, H129, H131, H132, H137, H85, 12, 13, 3, 7, HC103, HC105, HC53, HC55, HC74, HC81, HC83, HC86 for which expression data has been deposited as GSE3790 at Gene Expression Omnibus \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3790}. Column names in the expression data matrix contain the corresponding CEL file names.

The first 29 samples are control samples and the next 12 samples are disease samples.
}
\format{Numeric matrix with 23 rows and 41 columns.}
\source{\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3790}}
\keyword{datasets}
