\name{expression_GSE19380}
\docType{data}
\alias{expression_GSE19380}
\title{Expression data for GSE19380 (RNA mixture experiment)}
\description{
Numeric matrix containing normalized expression data for 31099 transcripts (rows) and 26 samples (columns).
}
\format{Numeric matrix with 31099 rows and 26 columns.}
\keyword{datasets}

