\name{groups}
\docType{data}
\alias{groups}
\title{Binary vector encoding sample group information.}
\description{
Numeric vector of length 41 containing 0s and 1s. 0s and 1s represent samples belonging to sample group 1 and 2, respectively.

This vector actually correspond to samples 1, 101, 11, 126, 14, 15, 17, 2, 20, 21, 52, 64, 8, H104, H109, H113, H115, H117, H118, H120, H121, H124, H126, H128, H129, H131, H132, H137, H85, 12, 13, 3, 7, HC103, HC105, HC53, HC55, HC74, HC81, HC83, HC86 for which expression data has been deposited as GSE3790 at Gene Expression Omnibus \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3790}. 0s are control samples  and 1s are disease samples.
}
\format{Numeric vector of length 41.}
\source{\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3790}}
\keyword{datasets}
