\name{marker}
\alias{marker}
\title{Reference signal}
\description{Generates a population-specific reference signal.}
\usage{marker(expr,id,sampleSubset=NULL,targetMean=1)}
\arguments{
  \item{expr}{Numeric matrix or ExpressionSet. The matrix contains the expression
	data with transcripts (i.e. for a microarray, probes) as rows and
	samples as columns. Row names should be defined and are used to determine
	the transcripts used to generate the reference signal (specified with the
	argument id). If an ExpressionSet is provided the expression data
	is extracted with the function exprs.}
  \item{id}{List of strings. Names of the transcripts to use to generate the
	reference signal. Names correspond to row names in expr if it is a matrix or
	row names in exprs(expr) if it is an ExpressionSet).}
  \item{sampleSubset}{Integer vector. Represents a subset of samples (specified
	as column indices in expr) to use to normalize the reference signal
        (see Details). By default all samples are used.}
  \item{targetMean}{Numeric. Mean of the reference signal (defaults to 1).}
}
\details{
A reference signal for a given population is generated by averaging the 
expression of multiple transcripts that are expressed in a cell 
population-specific manner.

The reference signal is scaled to have an average value determined by targetMean
(1 by default). The argument sampleSubset restricts the samples to take into
account for normalization (i.e. the average of the reference signal over 
sampleSubset will be the value specified by targetMean, but the reference signal
will still be of length equal to the number of columns in expr).

The expression value of each transcript is normalized before averaging over
multiple transcripts so that they have same weight in the final reference signal
(i.e to avoid that highly expressed transcripts dominate the reference signal).

The names of transcripts to be used are specified as items of a list (id). If
the expression matrix contains multiple measures of the same transcript (for 
instance several probes of a microarray measuring the same transcript) they can 
be averaged before taking the average over different transcripts. Thereby every 
transcript has the same weight in the final reference signal, independently of 
the number of probes used to measure it.  Multiple measures of the same 
transcript are specified as a vector in the list.
}
\value{
  \item{refSignal}{Numeric vector of the same length as the number of columns in
expr (or in exprs(expr) if expr is provided as an ExpressionSet). Reference signal
(see Details).}
}
\references{Kuhn A, Thu D, Waldvogel HJ, Faull RL, Luthi-Carter R. Population-specific expression analysis (PSEA) reveals molecular changes in diseased brain. Nat Methods 2011, 8(11):945-7}
\author{Alexandre Kuhn \email{alexandre.m.kuhn@gmail.com}}
\examples{
## Load example expression data
data("example")

## List of probes corresponding to neuronal genes:
## 221805_at, 221801_x_at, 221916_at all measure NEFL expression.
## 201313_at, 210040_at, 205737_at, 210432_s_at measure
## ENO2, SLC12A5, KCNQ2 and SCN3A expression, respectively.
neuron_probesets <- list(c("221805_at", "221801_x_at", "221916_at"),
		"201313_at", "210040_at", "205737_at", "210432_s_at")

## Neuronal reference signal
neuron_reference <- marker(expression, neuron_probesets)
}
